backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.31.1
python-multipart>=0.0.19

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiomysql==0.3.0
alembic==1.13.1

# MCP
mcp==1.23.0

# Gemini (2단계)
google-generativeai==0.8.0

# Utilities
pydantic[email]>=2.8.0
email-validator>=2.0.0
pydantic-settings>=2.4.0
python-dotenv==1.0.0
httpx>=0.27.0

# Security
cryptography>=44.0.1
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3